sokovnik v4

+++ src/sokovnik/Jabuka.java
@@ -6,7 +6,7 @@ java.util.*;
 public class Jabuka extends Vocka {
 	private String sortaJabuke;
 	
-	public Jabuka(String nazivVocke, double tezinaVocke, boolean isTrula) {
+	public Jabuka(String nazivVocke, int tezinaVocke, boolean isTrula) {
 		super(nazivVocke, tezinaVocke, isTrula);
 		setSortaJabuke();
 	}
@@ -14,7 +14,7 @@ public class Jabuka extends Vocka {
 	public void setSortaJabuke() {
 		String[] sorte = {"Zlatni delises", "Crveni delises", "Greni smit"};
 		Random rand = new Random();
-        this.sortaJabuke = sorte[rand.nextInt(sorte.length)];
+        this.sortaJabuke=  sorte[rand.nextInt(sorte.length)];
 	}
 	
 	public String getSorta() {
